Ship the TUI as `comet tui`, not a separate cometui command
Replaces the two-installer cometui approach: the viewport is now a subcommand
of the one binary. `comet tui` attaches to a running engine or spawns one —
and because it lives in the `comet` binary, the engine it spawns is always
right there. No second binary to bundle, no standalone installer, no name that
collides with the desktop app.

The run logic (arg parsing, env resolution, file tracing, current-thread
runtime) moves into `comet_tui::cli` so the shipped `comet tui` and the
`comet-tui` dev binary share one entry point and can't drift. `comet`'s main
skips its stdout tracing for the tui path — a log line mid-frame would corrupt
the alternate screen.

Reverts the cometui packaging/installer/edge-route/release-job added in the
previous commit; the `comet` package already carries the viewport now.
