+++ .github/workflows/release.yml
@@ -68,36 +68,9 @@ jobs:
           path: target/package/*.dmg
           if-no-files-found: error
 
-  # The `cometui` terminal binary — a peer of the desktop app, not part of it.
-  # Its own job because it has no gpui dependency: it needs none of the display
-  # libraries the `comet` builds do, and it must ship even when the flaky macOS
-  # gpui build (above) fails. Same `v<version>` tag, same R2 bucket, served by
-  # `curl … /install-tui.sh`.
-  tui:
-    strategy:
-      fail-fast: false
-      matrix:
-        include:
-          - runner: ubuntu-24.04 # linux x86_64
-          - runner: ubuntu-24.04-arm # linux aarch64
-          - runner: macos-latest # macos arm64
-    runs-on: ${{ matrix.runner }}
-    steps:
-      - uses: actions/checkout@v4
-      - name: rust toolchain (rust-toolchain.toml)
-        run: rustup show active-toolchain || rustup toolchain install
-      - uses: Swatinem/rust-cache@v2
-      - name: package
-        run: scripts/package-tui.sh
-      - uses: actions/upload-artifact@v4
-        with:
-          name: tui-${{ runner.os }}-${{ runner.arch }}
-          path: target/package/comet-tui-*.tar.gz
-          if-no-files-found: error
-
   publish:
     if: startsWith(github.ref, 'refs/tags/v')
-    needs: [linux, macos, tui]
+    needs: [linux, macos]
     runs-on: ubuntu-latest
     env:
       # `secrets` is not a valid context in step `if:` expressions — surface
