+++ apps/comet/Cargo.toml
@@ -12,6 +12,7 @@ path = "src/main.rs"
 [dependencies]
 comet-engine.workspace = true
 comet-ui.workspace = true
+comet-tui.workspace = true
 tokio.workspace = true
 clap.workspace = true
 anyhow.workspace = true
