+++ apps/comet/src/main.rs
@@ -29,6 +29,9 @@ enum Command {
         #[command(subcommand)]
         command: DaemonCommand,
     },
+    /// Terminal viewport over the same engine — attaches to a running app or
+    /// daemon, or starts one, and detaches (leaving work running) when it exits.
+    Tui(comet_tui::cli::TuiArgs),
 }
 
 #[derive(Subcommand)]
@@ -79,20 +82,25 @@ fn workos_client_id_from_env(edge_token: &Option<String>) -> Option<String> {
 
 fn main() -> anyhow::Result<()> {
     let cli = Cli::parse();
-    // Long-running modes log at info; the one-shot CLI commands keep their
-    // stdout clean (RUST_LOG still overrides either default).
-    let default_filter = match &cli.command {
-        None | Some(Command::Headless) => "info",
-        Some(_) => "warn",
-    };
-    tracing_subscriber::fmt()
-        .with_env_filter(
-            tracing_subscriber::EnvFilter::try_from_default_env()
-                .unwrap_or_else(|_| default_filter.into()),
-        )
-        .init();
+    // The TUI owns its own tracing (to a file — a line on stdout would land
+    // inside the alternate screen and corrupt it), so skip the global stdout
+    // subscriber entirely for it. Everything else logs to stdout: long-running
+    // modes at info, one-shot CLI commands at warn (RUST_LOG overrides either).
+    if !matches!(cli.command, Some(Command::Tui(_))) {
+        let default_filter = match &cli.command {
+            None | Some(Command::Headless) => "info",
+            Some(_) => "warn",
+        };
+        tracing_subscriber::fmt()
+            .with_env_filter(
+                tracing_subscriber::EnvFilter::try_from_default_env()
+                    .unwrap_or_else(|_| default_filter.into()),
+            )
+            .init();
+    }
 
     match cli.command {
+        Some(Command::Tui(args)) => comet_tui::cli::run(args),
         Some(Command::Headless) => {
             let runtime = tokio::runtime::Runtime::new()?;
             runtime.block_on(async {
