+++ apps/tui/Cargo.toml
@@ -11,7 +11,5 @@ path = "src/main.rs"
 
 [dependencies]
 comet-tui.workspace = true
-tokio.workspace = true
 clap.workspace = true
 anyhow.workspace = true
-tracing-subscriber.workspace = true
